pyproject.toml
[project]
name = "docforge"
version = "0.1.0"
description = "Universal document parser for LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10", "Pillow>=10.0.0"]
easyocr = ["easyocr>=1.7.0"]
docx = ["python-docx>=1.0.0"]
html = ["beautifulsoup4>=4.12.0", "readability-lxml>=0.8.0"]
email = ["extract-msg>=0.48.0"]
tables = ["img2table>=1.2.0"]
all = ["docforge[ocr,easyocr,docx,html,email,tables]"]
dev = [
    "pytest>=8.0.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "fpdf2>=2.7.0",
]

[project.scripts]
docforge = "docforge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
